Ticket #953 - Support of automatic language parameter in URLS.
static and param elements in urls.xml can now have a type attribute with 'lang' or 'locale' value.
For param elements of these types, there is no need to give them when calling jUrl::get.>
At startup, the language in jApp::config()->locale is set automatically with the value of the param or static element that have a type 'lang' or 'locale'.