Improved support of locale/lang.
We don't add same value as lang for country when country doesn't exist, since it has no sens for most of offical locales. new methods getCorrespondingLocale, getPreferedLocaleFromRequest and langToLocale on jLocale.
The configuration parameter availableLanguageCode of the autolocale plugin has been moved to the main configuration and renamed to availableLocales.