Removed the pivot class generated for user dao record
It has a complicated name. It causes issues when we have the same dao used with two db driver.
And finally, it does not follow the 'spirit'. Since we 'import', we don't have an inheriting behavior on
the user record. And probably the creator of the user record don't want inheriting.