Missing parameter for createFile
The 3th parameter is missing for the createFile function. This array is useless for now (the template "module/template.tpl" doesn't need it), but generate a php warning when the createzone is used. I choosed to keep the $param variable instead of an empty array in cases of futures needs.