jauthdb_admin should be accessible with an other driver than db
An application could provide a jauth driver that is not "Db", but
which uses a Dao. So we should be able to use jauthdb_admin
to manage with this driver since it has a dao.

In the configuration of the driver, a parameter compatiblewithdb=on
should be set to certifie that jauthdb_admin can manage users
with this driver.