jForms: group control can now have a checkbox
The checkbox enable/disable the content of the group.
To have this feature, the withcheckbox attribute should be set on the
<group> element, with the "true" value.

In this case, <group> supports child elements <oncheckvalue> and <onuncheckvalue>.
It supports also the defaultvalue attribute which content on of two values of
the checkbox state ('0' or '1' by default).

The class jforms-notselected is added on the fieldset when the checkbox is
unchecked.

The fieldset has now the jforms-ctrl-group class.