Fixed issues on jEvents tests
The cache file of jevents had a bad name and never the same name for
each tests.

Add the possibility to force the loading of cache files generated
by jIncluder. So, even the cache file has been loaded one time,
it can be read again. Usefull for tests that erase data created
by a cache file.

Closes #226