Significant url engine can parse/generate basic sign. url
It now ignores 'module' and 'action' parameters into
the url to parse, and generates always an url like
"/<module>/<controller>/<action>" when no url are defined
for an action in urls.xml.

Parsing "/<module>/<controller>/<action>" fails if there is
no defined url and there are no dedicated entrypoint for
the given module.