[NEW] Implements web assets system
It allows to define groups of CSS and JS to include in the response.
It is defined in the configuration, not in the code, so you
can change assets needed by a module easily.
Example of use: change the path of a js lib, overriding a css file,
define different assets for dev and prod environment...