Upgrade jQuery-UI to 1.12.1
Since jQuery-UI does not provides separate files for
each widget, we provide from now a build of jQueryUI
containing only the datepicker widget, because the default
jQuery-UI is too heavy (500KB!!) for a datepicker.

If the user want a different version of jQueyr with more
features, he has to put files of jQuery into its app
and change the webassets configuration.