Backup uninstaller script when unconfigure
After unconfigure a module, we may remove the source code of the
module. But then, when deploying the application and launch the
installer, the uninstaller script of removed module cannot be
executed. When unconfiguring a module, we should then backup
the uninstaller script into the application.

This script and other needed information to uninstall
the module in an application instance, can be removed
when we are sure that all application instance
have been migrated.