jForms: new control for uploads with enhanced management
This control and its associated widget allow the user to choose in the form
whether he keeps the existing file, or he uploads a new file, or he
deletes the existing file (when the control is not required).

To use this new control, use the <upload> tag with the attribute
controlclass="jFormsControlUpload2".