Move jDB tools classes from outside drivers
Since jDb tools classes are more about an SQL language than a
database api, it does make more sens to allow to access to them
from outside a driver.

It allows too to fix an issue when trying to compile a DAO class
during the load of a session. In this case, we only have the db type,
not the driver, so we cannot to load a driver in order to access to
a db tools class.