[jzone] fix generation of the cache key, when parameters contains unwanted values
some parameters may have values (like some object properties) that are
not used for the zone and unfortunately which may changed often. The
Cache key is then not 'stable' and the cache system may generate a
cache at each call.

A new method getCacheId() can be redefined to return only parameter
values that determines the uniqueness of the zone content.