Fix configuration cache when the application is accessible with several domains
The cache was created with the first domain used to access to the application.
So for other domains (like a new domain name), jRequest::getDomainName() always
returned the first domain, which may be not valid after some time.

The issue was then on generating full urls: they have the same first domain, because
it relies on informations stored into the cache file.

The cache file should be regenerate if the domain change. To ease the process, the name
of the cache file contains now the domain name.