Fix installer: useDbProfile must not change the profile name
It should not check if there is an alias, jProfiles is already doing
the job. jProfiles takes in priority the section of a profile, before
to check if there is an alias, whereas the installer did the contrary.
So when a profile was declared both as an alias and as a section,
tables may not be installed in the same database as the one indicated
into the profile section.