Fix error during postgresql connection close
When profiles has the same connection parameters, the pgsql
extension use the same internal connection. So when each jDb connection
are closed, the pg_close fails on the second jDb connection, because
the internal pg connection is already closed.

All profiles having the same connection parameters and same
settings for pgsql, use the same jDb connection object. If settings
are differents, a new connection is used.