Configurator: fix error when a module and its dependencies are upgraded
There was an error about bad version of a dependency, when the module we want
to force to configure has been upgraded, and has some dependencies that are
upgraded too, and which have previous versions that don't match the new version
constraints indicated into the module.

The configurator process tried to compare the version constraints with the previous
version of the dependencies.

Now, when forcing to configure a module, its dependencies are also checked and version
constraints are compared to the new version of dependencies.