jForms: support of form submit with Xhr
The submit with Xhr in a jForms form is activated with the Javascript method
`submitWithXHR` of a jFormsJQForm object. some callback can be given to
this method: one in case the form is valid, the other in case there are errors.

These callbacks receive a JSON object built by the new jResponseFormJQJson response
object. The simplest usage of the response is to give the jFormsBase object before
returning the response.