Fix jForms: in JS, the form is now declared after the setup of all controls
It was declared just after the creation of the jforms JS object, and before
the initilization of controls.

So, starting from now, the event `jformsready` is triggered and the callbacks declared
with `onFormReady` are executed after the full initialization of the javascript
objects of jForms.

If you want to keep the old behavior, you should indicate
the option `'deprecatedDeclareFormBeforeControls'=>true` to the form builder.