[tests] improve the dev stack for debugging
Xdebug is configured into the phpfpm container.

Docker compose can now be launched without using run-docker, if the .env file
is set correctly.

The docker-compose.yml contains the project name. So if your IDE is playing
with one of the container, you may set the COMPOSE_PROJECT_NAME environment
variable.